refactor(server): carve out the workspace API
Workspace and project behavior was the largest remaining domain coupled directly to the server composition root, which kept Git-heavy tests and runtime lifecycle concerns entangled with unrelated APIs. Give the domain a concrete handler boundary so its routes, observers, cache, terminals, and project operations can evolve and test together while the root retains process-wide orchestration.\n\nThe shared diff contract lives below both packages to preserve existing Huma schema identities, and dependent Kata and Fleet integrations consume narrow workspace contracts instead of compatibility wrappers.
