+++ internal/server/fleet_worktree_links.go
@@ -281,8 +281,8 @@ func (s *Server) NotifyWorktreeLinksChanged() {
 // notifyWorktreeStatsChanged is the stats-sampler callback: it
 // broadcasts a payload-free worktree_stats_changed refetch hint.
 func (s *Server) notifyWorktreeStatsChanged() {
-	if s.workspaceDiffCache != nil {
-		s.workspaceDiffCache.RevalidateSelected()
+	if s.workspaceAPI != nil {
+		s.workspaceAPI.RevalidateSelectedDiffs()
 	}
 	if s.hub != nil {
 		s.hub.Broadcast(Event{
