+++ internal/server/workspaceapi/projects_clone.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
