+++ internal/server/workspaceapi/projects_inspect_branches.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
