+++ internal/server/workspaceapi/tmux_activity_test.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
