+++ internal/server/workspaceapi/workspace_diff_cache.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
