+++ internal/server/workspaceapi/workspace_diff_cache_test.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
