+++ internal/server/workspaceapi/workspace_runtime_terminal_test.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
