+++ internal/server/workspaceapi/worktree_stale.go
@@ -1,4 +1,4 @@
-package server
+package workspaceapi
 
 import (
 	"context"
