--- internal/server/ws_debug.go
@@ -1,23 +1,0 @@
-package server
-
-import (
-	"log/slog"
-	"os"
-	"strings"
-)
-
-func logWebsocketDebug(msg string, args ...any) {
-	if !websocketDebugEnabled() {
-		return
-	}
-	slog.Debug(msg, args...)
-}
-
-func websocketDebugEnabled() bool {
-	switch strings.ToLower(strings.TrimSpace(os.Getenv("MIDDLEMAN_WS_DEBUG"))) {
-	case "1", "true", "yes", "on":
-		return true
-	default:
-		return false
-	}
-}
